@@ -3,22 +3,22 @@ package com.mrbysco.telepass.item;
 import com.mrbysco.telepass.Reference;
 import com.mrbysco.telepass.init.TeleGroup;
 import com.mrbysco.telepass.util.PlayerUtil;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.util.ActionResult;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.Hand;
-import net.minecraft.util.SoundCategory;
-import net.minecraft.util.SoundEvents;
-import net.minecraft.util.Util;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.StringTextComponent;
-import net.minecraft.util.text.TextFormatting;
-import net.minecraft.util.text.TranslationTextComponent;
-import net.minecraft.world.World;
+import net.minecraft.ChatFormatting;
+import net.minecraft.Util;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
 import net.minecraftforge.common.util.FakePlayer;
 
 public class ItemTeleCompass extends Item {
@@ -35,57 +35,56 @@ public class ItemTeleCompass extends Item {
 	}
 
 	@Override
-	public ActionResult<ItemStack> use(World worldIn, PlayerEntity playerIn, Hand handIn) {
+	public InteractionResultHolder<ItemStack> use(Level worldIn, Player playerIn, InteractionHand handIn) {
 		ItemStack itemstack = playerIn.getItemInHand(handIn);
-		
-		if(!worldIn.isClientSide && itemstack.hasTag() && itemstack.getTag().contains(Reference.OWNER_TAG)) {
+
+		if (!worldIn.isClientSide && itemstack.hasTag() && itemstack.getTag().contains(Reference.OWNER_TAG)) {
 			String ownerName = itemstack.getTag().getString(Reference.OWNER_TAG);
-			if(ownerName.equalsIgnoreCase(playerIn.getName().getContents())) {
-        		playerIn.sendMessage(new TranslationTextComponent("item.telepass.self"), Util.NIL_UUID);
-        		return new ActionResult<>(ActionResultType.SUCCESS, itemstack);
+			if (ownerName.equalsIgnoreCase(playerIn.getName().getContents())) {
+				playerIn.sendMessage(new TranslatableComponent("item.telepass.self"), Util.NIL_UUID);
+				return new InteractionResultHolder<>(InteractionResult.SUCCESS, itemstack);
 			}
 			boolean isOnline = false;
 
-			for(PlayerEntity player : worldIn.players()) {
-				if(player.getName().getContents().equalsIgnoreCase(ownerName)) {
+			for (Player player : worldIn.players()) {
+				if (player.getName().getContents().equalsIgnoreCase(ownerName)) {
 					isOnline = true;
 					break;
 				}
 			}
 
-			if(isOnline) {
-				PlayerEntity owner = PlayerUtil.getPlayerEntityByName(worldIn, ownerName);
-				if(owner != null && owner.getCommandSenderWorld().dimension().getRegistryName() != playerIn.getCommandSenderWorld().dimension().getRegistryName()) {
-					playerIn.sendMessage(new TranslationTextComponent("item.telepass.dimension", TextFormatting.RED + ownerName), Util.NIL_UUID);
-					return new ActionResult<>(ActionResultType.SUCCESS, itemstack);
+			if (isOnline) {
+				Player owner = PlayerUtil.getPlayerEntityByName(worldIn, ownerName);
+				if (owner != null && owner.getCommandSenderWorld().dimension().getRegistryName() != playerIn.getCommandSenderWorld().dimension().getRegistryName()) {
+					playerIn.sendMessage(new TranslatableComponent("item.telepass.dimension", ChatFormatting.RED + ownerName), Util.NIL_UUID);
+					return new InteractionResultHolder<>(InteractionResult.SUCCESS, itemstack);
 				}
 
-				if (!playerIn.abilities.instabuild) {
+				if (!playerIn.getAbilities().instabuild) {
 					itemstack.hurtAndBreak(1, playerIn, (p_219998_1_) -> p_219998_1_.broadcastBreakEvent(handIn));
 				}
 
-				if(!(playerIn instanceof FakePlayer) && owner != null) {
-					worldIn.playSound((PlayerEntity)null, playerIn.getX(), playerIn.getY(), playerIn.getZ(), SoundEvents.CHORUS_FRUIT_TELEPORT, SoundCategory.NEUTRAL, 0.5F, 0.4F / (random.nextFloat() * 0.4F + 0.8F));
+				if (!(playerIn instanceof FakePlayer) && owner != null) {
+					worldIn.playSound((Player) null, playerIn.getX(), playerIn.getY(), playerIn.getZ(), SoundEvents.CHORUS_FRUIT_TELEPORT, SoundSource.NEUTRAL, 0.5F, 0.4F / (worldIn.random.nextFloat() * 0.4F + 0.8F));
 					playerIn.getCooldowns().addCooldown(this, 20);
 
 					playerIn.teleportTo(owner.getX(), owner.getY(), owner.getZ());
 				}
 			} else {
-				playerIn.sendMessage(new TranslationTextComponent("item.telepass.offline", TextFormatting.RED + ownerName), Util.NIL_UUID);
+				playerIn.sendMessage(new TranslatableComponent("item.telepass.offline", ChatFormatting.RED + ownerName), Util.NIL_UUID);
 			}
-			return new ActionResult<>(ActionResultType.SUCCESS, itemstack);
+			return new InteractionResultHolder<>(InteractionResult.SUCCESS, itemstack);
 		} else {
-			return new ActionResult<>(ActionResultType.FAIL, itemstack);
+			return new InteractionResultHolder<>(InteractionResult.FAIL, itemstack);
 		}
 	}
 
 	@Override
-	public void inventoryTick(ItemStack stack, World worldIn, Entity entityIn, int itemSlot, boolean isSelected) {
-		if(!worldIn.isClientSide) {
-			if((stack.hasTag() && stack.getTag() != null && !stack.getTag().contains(Reference.OWNER_TAG)) || !stack.hasTag() || stack.getTag() == null) {
-				CompoundNBT tag = stack.getTag() == null ? new CompoundNBT() : stack.getTag();
-				if(entityIn instanceof PlayerEntity && !(entityIn instanceof FakePlayer)) {
-					PlayerEntity player = (PlayerEntity) entityIn;
+	public void inventoryTick(ItemStack stack, Level worldIn, Entity entityIn, int itemSlot, boolean isSelected) {
+		if (!worldIn.isClientSide) {
+			if ((stack.hasTag() && stack.getTag() != null && !stack.getTag().contains(Reference.OWNER_TAG)) || !stack.hasTag() || stack.getTag() == null) {
+				CompoundTag tag = stack.getTag() == null ? new CompoundTag() : stack.getTag();
+				if (entityIn instanceof Player player && !(entityIn instanceof FakePlayer)) {
 					tag.putString(Reference.OWNER_TAG, player.getName().getContents());
 
 					stack.setTag(tag);
@@ -96,11 +95,11 @@ public class ItemTeleCompass extends Item {
 	}
 
 	@Override
-	public ITextComponent getName(ItemStack stack) {
-		if(stack.hasTag() && stack.getTag() != null && stack.getTag().contains(Reference.OWNER_TAG)) {
-			CompoundNBT tag = stack.getTag();
+	public Component getName(ItemStack stack) {
+		if (stack.hasTag() && stack.getTag() != null && stack.getTag().contains(Reference.OWNER_TAG)) {
+			CompoundTag tag = stack.getTag();
 			String owner = tag.getString(Reference.OWNER_TAG);
-			return new StringTextComponent(owner + "'s ").append(new TranslationTextComponent(this.getDescriptionId(stack)));
+			return new TextComponent(owner + "'s ").append(new TranslatableComponent(this.getDescriptionId(stack)));
 		} else {
 			return super.getName(stack);
 		}
