@@ -4,18 +4,18 @@ import com.mrbysco.telepass.TelePass;
 import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.common.ForgeConfigSpec.IntValue;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.config.ModConfig;
+import net.minecraftforge.fml.event.config.ModConfigEvent;
 import org.apache.commons.lang3.tuple.Pair;
 
 public class TeleConfig {
 
-	public static class Server {
+	public static class Common {
 		public final IntValue goldDurability;
 		public final IntValue diamondDurability;
 
-		Server(ForgeConfigSpec.Builder builder) {
-			builder.comment("Server settings")
-					.push("Server");
+		Common(ForgeConfigSpec.Builder builder) {
+			builder.comment("General settings")
+					.push("general");
 
 			goldDurability = builder
 					.comment("Defines the amount of uses the Gold TelePass has [Default: 15]")
@@ -28,22 +28,23 @@ public class TeleConfig {
 			builder.pop();
 		}
 	}
-	public static final ForgeConfigSpec serverSpec;
-	public static final Server SERVER;
+
+	public static final ForgeConfigSpec commonSpec;
+	public static final Common COMMON;
 
 	static {
-		final Pair<Server, ForgeConfigSpec> specPair = new ForgeConfigSpec.Builder().configure(Server::new);
-		serverSpec = specPair.getRight();
-		SERVER = specPair.getLeft();
+		final Pair<Common, ForgeConfigSpec> specPair = new ForgeConfigSpec.Builder().configure(Common::new);
+		commonSpec = specPair.getRight();
+		COMMON = specPair.getLeft();
 	}
 
 	@SubscribeEvent
-	public static void onLoad(final ModConfig.Loading configEvent) {
-		TelePass.logger.debug("Loaded TelePass' config file {}", configEvent.getConfig().getFileName());
+	public static void onLoad(final ModConfigEvent.Loading configEvent) {
+		TelePass.LOGGER.debug("Loaded TelePass' config file {}", configEvent.getConfig().getFileName());
 	}
 
 	@SubscribeEvent
-	public static void onFileChange(final ModConfig.Reloading configEvent) {
-		TelePass.logger.debug("TelePass' config just got changed on the file system!");
+	public static void onFileChange(final ModConfigEvent.Reloading configEvent) {
+		TelePass.LOGGER.warn("TelePass' config just got changed on the file system!");
 	}
 }
