@@ -5,16 +5,16 @@ import com.mrbysco.telepass.config.TeleConfig;
 import java.util.function.Supplier;
 
 public enum CompassMaterial {
-    GOLD(() -> TeleConfig.SERVER.goldDurability.get()),
-    DIAMOND(() -> TeleConfig.SERVER.diamondDurability.get());
+	GOLD(() -> TeleConfig.COMMON.goldDurability.get()),
+	DIAMOND(() -> TeleConfig.COMMON.diamondDurability.get());
 
-    private final Supplier<Integer> durability;
+	private final Supplier<Integer> durability;
 
-    CompassMaterial(Supplier<Integer> durability) {
-        this.durability = durability;
-    }
+	CompassMaterial(Supplier<Integer> durability) {
+		this.durability = durability;
+	}
 
-    public int getMaxUses() {
-        return this.durability.get();
-    }
+	public int getMaxUses() {
+		return this.durability.get();
+	}
 }
